+++ Blazor_Server/Model/IIdeaManager.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace Blazor_Server.Data
+namespace Blazor_Server.Model
 {
     public interface IIdeaManager
     {
