+++ Blazor_Server/Model/IProjectFileManager.cs
@@ -2,7 +2,7 @@
 using System.Security.Principal;
 using System.Threading.Tasks;
 
-namespace Blazor_Server.Data
+namespace Blazor_Server.Model
 {
     public interface IProjectFileManager
     {
