+++ Blazor_Server/Model/ProjectFile.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components.Forms;
 using System;
 
-namespace Blazor_Server.Data
+namespace Blazor_Server.Model
 {
     public class ProjectFile : IProjectFile
     {
