+++ Blazor_Server/Pages/ProjectList.razor
@@ -1,4 +1,5 @@
 ﻿@using Blazor_Server.Data
+@using Blazor_Server.Model
 @inject ProjectManagerService ProjectManager
 
 @page "/projectlist"
