+++ Blazor_Server/Shared/Idea.razor
@@ -1,4 +1,6 @@
 ﻿@using Blazor_Server.Data
+@using Blazor_Server.Model
+
 @inject ProjectManagerService ProjectManager
 @inject NavigationManager NavigationManager
