+++ Blazor_Server/Shared/Table.razor
@@ -1,4 +1,4 @@
-﻿@using Blazor_Server.Data
+﻿@using Blazor_Server.Attributes
 @using System.ComponentModel
 @using System.Reflection
